Add Docker and GitHub workflow for image & deployment
This commit introduces a Dockerfile to containerize the application.
It also features GitHub workflows for building a Docker image, pushing it to the GitHub container registry, and deploying to a Kubernetes cluster.
The changes were necessary for facilitating continuous integration and deployment pipelines.

Also introduced are dependabot configuration for cargo & github-actions ecosystem and .dockerignore & .gitattributes for docker & git specific rules. Version updates in toml files were made to keep dependencies up to date.