Update gif handling and simplify response sending
The changes in this commit are mainly about refactoring the way the program handles gifs and simplifying the process of sending the response to the gifs.

Firstly, the version was updated to 0.1.6 in Cargo.toml, signaling a new release.

As to functional changes, in various places `.unwrap_or()` method has been used, effectively eliminating the need for separate blocks dealing with certain situations, thus reducing redundancy and improving the efficiency of the program.

A new messaging module was created, featuring a function to send replies. This helps in having a single place for all related code, making it easy to maintain and fix in the future. It checks for messages and sends available messages asynchronously. This also allows processing multiple messages and displaying them using a single command "send_reply".

In the meantime, we reduced unneeded imports, removed unneeded variables, and improved the 'game' command execution, changing the variable name from 'game_query' to 'game' to better reflect its purpose. Fields visibility has also been updated to meet minimum requirements.

Overall, these changes aim to simplify and optimize the code structure, making it more straightforward and easy to maintain, while still providing the same functionality.